Delete the now-defunct BagFile type
