--- common/src/main/scala/uk/ac/wellcome/platform/archive/common/bagit/models/BagFile.scala
@@ -1,8 +1,0 @@
-package uk.ac.wellcome.platform.archive.common.bagit.models
-
-import uk.ac.wellcome.platform.archive.common.verify.Checksum
-
-case class BagFile(
-  checksum: Checksum,
-  path: BagPath
-)
