+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/fixtures/StorageRandomThings.scala
@@ -175,4 +175,12 @@ trait StorageRandomThings extends RandomThings {
 
     algorithms(Random.nextInt(algorithms.length))
   }
+
+  def createBagPath: BagPath = BagPath(randomAlphanumeric)
+
+  def createChecksumWith(algorithm: HashingAlgorithm = SHA256): Checksum =
+    Checksum(algorithm = algorithm, value = randomChecksumValue)
+
+  def createChecksum: Checksum =
+    createChecksumWith()
 }
