+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/storage/services/StorageManifestServiceTest.scala
@@ -30,7 +30,6 @@ class StorageManifestServiceTest
     extends FunSpec
     with Matchers
     with BagGenerators
-    with BagFileGenerators
     with FetchMetadataGenerators
     with StorageLocationGenerators
     with StorageSpaceGenerators
